Landingpage integration
In order to make it possible for users to register/login the application is separated from the usual fmh template.
The application gets its own page structure with a header different from the fmh site.
In this commit the basic modals are implemented and the logic for showing the base template was adjusted.